refactor: moves utils.py to helpers
