+++ eolic/meta/__init__.py
@@ -1,0 +1,1 @@
+"""Module containing metaclasses."""
